Moved Contributions.ts to frontend/config
- Moved contributions to frontend/config & started refactoring code / redesigning schema in preperation to move the types to the backend under contributors.model.ts
- Removed unused import
